Remove virtual service sync isolation rules
