+++ src/meson.build
@@ -32,13 +32,13 @@ dp_sources = [
   'nodes/rx_periodic_node.c',
   'nodes/snat_node.c',
   'nodes/tx_node.c',
-  'rte_flow/dp_rte_flow.c',
-  'rte_flow/dp_rte_flow_capture.c',
-  'rte_flow/dp_rte_flow_init.c',
-  'rte_flow/dp_rte_flow_traffic_forward.c',
   'rte_flow/dp_rte_async_flow.c',
   'rte_flow/dp_rte_async_flow_isolation.c',
   'rte_flow/dp_rte_async_flow_template.c',
+  'rte_flow/dp_rte_flow.c',
+  'rte_flow/dp_rte_flow_capture.c',
+  'rte_flow/dp_rte_flow_isolation.c',
+  'rte_flow/dp_rte_flow_traffic_forward.c',
   'dp_argparse.c',
   'dp_cntrack.c',
   'dp_conf.c',
